requirements.txt
flask
phidata
python-dotenv
duckduckgo-search
google-generativeai